Add files via upload
The issue lies (for me anyways) when the importing of the data and the manipulation are all within the same LOONG file. (As we have it now). Putting the data in a standalone file will not only increase readability but will also allow for me to import the necessary relevant data wherein we can then add on top of it.